+++ core/src/main/java/engineer/thesis/core/service/Implementation/UserService.java
@@ -3,6 +3,8 @@ package engineer.thesis.core.service.Implementation;
 import engineer.thesis.core.exception.AlreadyExistsException;
 import engineer.thesis.core.exception.TokenExpiredException;
 import engineer.thesis.core.model.*;
+import engineer.thesis.core.model.dto.RegisterOnBehalfRequest;
+import engineer.thesis.core.model.dto.RegisterRequest;
 import engineer.thesis.core.model.dto.ResetPasswordDTO;
 import engineer.thesis.core.model.dto.UserDTO;
 import engineer.thesis.core.repository.DoctorRepository;
@@ -10,8 +12,6 @@ import engineer.thesis.core.repository.PasswordResetTokenRepository;
 import engineer.thesis.core.repository.PatientRepository;
 import engineer.thesis.core.repository.UserRepository;
 import engineer.thesis.core.security.model.PasswordResetToken;
-import engineer.thesis.core.security.model.RegisterOnBehalfRequest;
-import engineer.thesis.core.security.model.RegisterRequest;
 import engineer.thesis.core.service.Interface.IUserService;
 import engineer.thesis.core.utils.CustomObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
