+++ core/src/main/java/engineer/thesis/core/service/Interface/IUserService.java
@@ -4,9 +4,9 @@ import engineer.thesis.core.exception.AlreadyExistsException;
 import engineer.thesis.core.exception.TokenExpiredException;
 import engineer.thesis.core.model.User;
 import engineer.thesis.core.model.UserRole;
+import engineer.thesis.core.model.dto.RegisterOnBehalfRequest;
+import engineer.thesis.core.model.dto.RegisterRequest;
 import engineer.thesis.core.model.dto.UserDTO;
-import engineer.thesis.core.security.model.RegisterOnBehalfRequest;
-import engineer.thesis.core.security.model.RegisterRequest;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
